GPU Codegen Expansion of Empty Memlets (#2006)
Memlets that the GPU code generator can not be expressed as a
`cudaMemcpy*()` call are turned into Maps.
However, there was an issue because codegen did not make sure that an
Memlet it wants to promote is empty.
So it accessed the subsets of an empty Memlets, which are `None` and
hits an error.