Positional Arguments for `CompiledSDFG` Restored From Cache (#2014)
When the `load_from_file()` function, which is used to load an SDFG from
the cache, then it did not specify the `argnames` argument of
`CompiledSDFG`'s constructor, with the effect that the `CompiledSDFG`
would not accept any positional arguments, although they where specified
and `sdfg._recompile = True; csdfg = sdfg.compile()` would accept them.

This PR also adds a warning if a `CompiledSDFG` was constructed this
way, but it is not an error, for backward compatibility reasons.