Make CUDACodeGenerator reentrant (#1995)
GT4Py-next is introducing support for parallel build of programs with
static arguments. In combination with the dace backend, this feature
results in multiple SDFG variants that differ for the specialization of
some static scalar arguments. By means of a thread pool, the SDFGs
should be optimized and compiled in parallel.

The global attribute `CUDACodeGen._in_device_code` currently makes the
dace backend non-reentrant. This PR makes `_in_device_code` an instance
attribute.

---------

Co-authored-by: Philip Müller <147368808+philip-paul-mueller@users.noreply.github.com>
Co-authored-by: alexnick83 <31545860+alexnick83@users.noreply.github.com>