Improve Validation  (#2083)
Improve validation to:
- Detect writes to scalars on interstate edges. For this I added
`_no_writes_to_scalars_or_arrays_on_interstate_edges` and a test to
`interstate_edge_utils_test` to check that an SDFG where we write to a
scalar on iedge is invalid.
- Do not mark tasklets with no incoming and outgoing memlets as sink
nodes. Accroding to discussion with Tal earlier the image from the SDFG
should be valid as it is not a sink node due to no incoming our outgoing
memlets, I improve validation to not mark this tasklet as sink node +
add a test case :
<img width="921" height="560" alt="image"
src="https://github.com/user-attachments/assets/9656de59-62bc-4f07-8116-050d6e06b749"
/>