Apply ConstantPropagation to AccessNodes (#2079)
`ConstantPropagation` pass used to work only for symbols/scalars in
expressions like `if` statements.
This PR extends the `replace_dict` functions which is called from the
`ConstantPropagation` to handle `AccessNode`s in case the `data` of an
`AccessNode` (if it's scalar) is passed in the replacement dictionary.
The `AccessNode` is replaced with a `tasklet` setting the constant
value.

---------

Co-authored-by: Philip Müller <philip.mueller@cscs.ch>
Co-authored-by: Tal Ben-Nun <tbennun@users.noreply.github.com>