Fix _get_assignments() (#2129)
**Description**

Passing: 

`*list(cfg.all_control_flow_blocks())`

as argument to the function

`cfg.all_edges(..)`

leads to a `KeyError`. The reason is, that
`cfg.all_control_flow_blocks()` returns `ControlFlowBlock`'s of
**nested** `ControlFlowRegion`'s which are not in `cfg._nodes` used by
the `all_edges(...)` function.

If I understand it correctly, the goal of the `_get_assignments(...)`
function is to find whether a symbol is written to
via an interstate edge. In this case, using `cfg.all_interstate_edges()`
should fix the issue.