Updated `Indices` (#2162)
This PR makes the `Indices` class a bit safer to use, a short summary:
- The `{max, min}_elements()` functions were returning a reference to
the underlying `self.indices` class, which might lead to an unintended
modification of the underlying data. The new solution is that they are
shallow copied.
- The constructor was changed, the class can now also be constructed
using single pass iterators.

This PR is also a RFC as I found some strange behaviour/cases, I marked
them with `Possible Error`.
So could somebody how knows a bit more about the code base take a look
at it and check if this behaviour is okay.