Loop Local Memory Reduction (#2160)
This PR adds the LoopLocalMemoryReduction transformation, also known as
K-Caching.
This transformation looks for loops, which access arrays with a fixed
alive window, then shrinks the array size to the window size by
introducing a modulo operation to the index computations (i.e. a
ringbuffer).

Example:
```py
for i in range(2, N):
        v[i] = a[i - 1] + a[i - 2]
        a[i] = b[i] * 2
```

turns into:
```py
for i in range(2, N):
        v[i] = a[(i - 1) % 3] + a[(i - 2) % 3]
        a[i % 3] = b[i] * 2
```

---------

Co-authored-by: Philipp Schaad <schaad.phil@gmail.com>