Added `canonicalize_memlet_trees()` (#2189)
In the DaCe of old [sic] a Memlet only had `.subset` and there was no
`.other_subset` and `.data` would always refer to what was on the
outside.
However, these days are gone, but as it can be seen for example in
[PR#2188](https://github.com/spcl/dace/pull/2188) not every part of the
code has been updated/changed/fixed to honor this brave new world.
The `canonicalize_memlet_trees()` function inspects a Map and the
Memlets that passing through them and modifies them such that they
conform to the old way.

---------

Co-authored-by: Philipp Schaad <schaad.phil@gmail.com>