Imporved `CompiledSDFG.fast_call()` (#2195)
The `fast_call()` method had to do an `import` statement to avoid a
circular import.
However, this import is only needed if a GPU check is needed/requested.

This PR relocates the `import` statement into the `if` branch where it
needed and by that avoid calling it if it is not needed, witch results
in a reduction of about 2us in the hot path.