Updated `invoke_compiled_sdfg_call_hooks` (#2196)
`invoke_compiled_sdfg_call_hooks` is used to add hooks at call time to
the `CompiledSDFG` object, for this `conextlib.contextmanager` is used,
to implement it as context manager.
This function is on the hot path and using `contextmanager` adds a non
negligible overhead, thus this PR implements the full context manager
object which is faster, especially in the case that no hook is
registered.