Deterministic `Map.label` for `MapFusion{Vertical, Horizontal}` (#2226)
Before the label of the Maps were not modified. Thus, the name of the
final Map dependent on the order in which they were fused. This PR
changes this by selecting the label in a deterministic way, i.e. always
use the "smaller" (`str.__lt__()` thinks).
Furthermore, it makes it explicit that `MapFusionHorizontal` preserves
the scope nodes of the first Map. This is done to make it more easy to
use in other contexts.

---------

Co-authored-by: Philipp Schaad <schaad.phil@gmail.com>